Clamp Project
This is a wooden clamp design. There are 5 parts. The top / bottom of the clamp are cut in a clamshell design, then glued / pinned together with brass rod. The pall / lever is s secondary part, I usually make them in batches by doing a pair with deep cuts. Then I can later separate it out into 4 pieces on a band saw. I usually make these parts from maple and use an ash bar. The bar should be a close fit, but still able to slide freely. You can also add sandpaper to the inside of the sliding foot for extra grip.